+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 1.5.0 -- 2019-05-30
+
+### Minor Changes
+* Add dependency on Apache Commons Codec 1.12.
+* Use org.apache.commons.codec.binary.Base64 instead of java.util.Base64 so
+  that the SDK can be used on systems that do not have java.util.Base64 but
+  support Java 8 language features.
+
+### Maintenance
+* Upgrade AWS Java SDK version from 1.11.169 to 1.11.561.
+* Upgrade Mockito from 2.23.4 to 2.28.1.
+* Upgrade Apache Commons Lang from 3.4 to 3.9.
+
 ## 1.4.1 -- 2019-05-10
 
 ### Patches